Move concrete class to impl package and add more tests
